// src/lib/api/schemas/updateUserRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * ciclo
 * OpenAPI spec version: 0.1.0
 */
import type { RoleId } from './roleId.ts';
import type { UserStatus } from './userStatus.ts';

export interface UpdateUserRequest {
  /** @nullable */
  email?: string | null;
  /** @nullable */
  father_last_name?: string | null;
  /** @nullable */
  mother_last_name?: string | null;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  password?: string | null;
  /** @nullable */
  role_ids?: RoleId[] | null;
  status?: null | UserStatus;
  /** @nullable */
  username?: string | null;
}